feat: defaultGridOptions
