+++ libs/handoverapp/src/lib/config/tableConfig.tsx
@@ -1,5 +1,6 @@
 import { HandoverPackage } from '@cc-components/handovershared';
-import { defaultColDef, statusColorMap } from '@cc-components/shared';
+import { statusColorMap } from '@cc-components/shared';
+import { defaultGridOptions } from '@cc-components/shared/workspace-config';
 import { DateCell, DescriptionCell, LinkCell, StatusCell } from '@cc-components/shared';
 import { BaseStatus } from '@cc-components/shared';
 
@@ -7,7 +8,7 @@ import { ICellRendererProps } from '@equinor/workspace-ag-grid';
 import { GridConfig } from '@equinor/workspace-fusion/grid';
 
 export const tableConfig: GridConfig<HandoverPackage> = {
-  gridOptions: { defaultColDef: defaultColDef },
+  gridOptions: defaultGridOptions,
   columnDefinitions: [
     {
       field: 'Comm pkg',
