+++ libs/shared/src/packages/table-helpers/src/lib/index.ts
@@ -1,2 +1,1 @@
 export * from './table';
-export * from './defaultColDef';
