+++ libs/shared/src/packages/workspace-config/src/index.ts
@@ -1,3 +1,4 @@
 export * from './useFilterConfig';
 export * from './useGridDataSource';
 export * from './useGardenDataSource';
+export * from './defaultGridOptions';
